IZE-254 fix signal handler
