+++ internal/commands/deploy/deploy.go
@@ -11,9 +11,9 @@ import (
 	"github.com/aws/aws-sdk-go/service/ssm"
 	"github.com/hazelops/ize/internal/apps"
 	"github.com/hazelops/ize/internal/config"
+	"github.com/hazelops/ize/internal/terraform"
 	"github.com/hazelops/ize/pkg/templates"
 	"github.com/hazelops/ize/pkg/terminal"
-	"github.com/hazelops/ize/pkg/terraform"
 	"github.com/pterm/pterm"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
