+++ internal/commands/destroy/destroy_infra.go
@@ -5,8 +5,8 @@ import (
 	"strings"
 
 	"github.com/hazelops/ize/internal/config"
+	"github.com/hazelops/ize/internal/terraform"
 	"github.com/hazelops/ize/pkg/terminal"
-	"github.com/hazelops/ize/pkg/terraform"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
