+++ internal/commands/terraform/terraform.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 
 	"github.com/hazelops/ize/internal/config"
+	"github.com/hazelops/ize/internal/terraform"
 	"github.com/hazelops/ize/pkg/templates"
-	"github.com/hazelops/ize/pkg/terraform"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
