+++ internal/docker/utils/signal_handlers.go
@@ -1,4 +1,4 @@
-package docker
+package utils
 
 import (
 	"context"
